% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge-mne.R
\name{hasMNE}
\alias{hasMNE}
\title{Is the Python \pkg{mne} module available?}
\usage{
hasMNE()
}
\value{
\code{TRUE} if \pkg{reticulate} and the Python \code{mne} module are
both available, otherwise \code{FALSE}.
}
\description{
Is the Python \pkg{mne} module available?
}
\examples{
hasMNE()
}
\seealso{
\code{\link[=pyMNEVersion]{pyMNEVersion()}}, \code{\link[=toMNE]{toMNE()}}, \code{\link[=fromMNE]{fromMNE()}}
}
