% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge-mne.R
\name{pyMNEVersion}
\alias{pyMNEVersion}
\title{Version of the available Python \pkg{mne} module}
\usage{
pyMNEVersion()
}
\value{
The \code{mne.__version__} string, or \code{NA_character_} when
\pkg{mne} is not available.
}
\description{
Version of the available Python \pkg{mne} module
}
\examples{
pyMNEVersion()
}
\seealso{
\code{\link[=hasMNE]{hasMNE()}}
}
